requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
tqdm>=4.65.0
yt-dlp>=2023.0.0
scipy>=1.11.0
